requirements.txt
opencv-python>=4.5.0
pyzmq>=22.0.0
numpy>=1.20.0
pytest>=7.0.0
